N-Player Pig
N-Player Pig is a multiplayer dice game that can be played by up to 6 people. Players race to reach 50 points. Each turn, a player repeatedly rolls a die until either a 1 is rolled or the player holds and scores the sum of the rolls. At any time during a player's turn, the player is faced with two decisions: 

1. Roll :
    - If the player rolls a 1, the player scores nothing and it becomes the opponent's turn. 
    - If the player rolls a 2-6, the number is added to the player's turn total and the player's turn continues. 

2. Hold :
    - The turn total is added to the player's score and it becomes the opponent's turn.